requirements.txt
requests
click
beautifulsoup4
cfscrapegophish
